Rename snapshoter to snapshotter
