+++ src/ggrc/snapshotter/datastructures.py
